IGNITE-19440 RebalanceUtil refactoring.
