+++ modules/table/src/main/java/org/apache/ignite/internal/utils/RebalanceUtil.java
@@ -18,14 +18,12 @@
 package org.apache.ignite.internal.utils;
 
 import static org.apache.ignite.internal.metastorage.dsl.Conditions.and;
-import static org.apache.ignite.internal.metastorage.dsl.Conditions.exists;
 import static org.apache.ignite.internal.metastorage.dsl.Conditions.notExists;
 import static org.apache.ignite.internal.metastorage.dsl.Conditions.or;
 import static org.apache.ignite.internal.metastorage.dsl.Conditions.revision;
 import static org.apache.ignite.internal.metastorage.dsl.Conditions.value;
 import static org.apache.ignite.internal.metastorage.dsl.Operations.ops;
 import static org.apache.ignite.internal.metastorage.dsl.Operations.put;
-import static org.apache.ignite.internal.metastorage.dsl.Operations.remove;
 import static org.apache.ignite.internal.metastorage.dsl.Statements.iif;
 
 import java.nio.charset.StandardCharsets;
@@ -37,185 +35,20 @@ import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 import org.apache.ignite.internal.affinity.AffinityUtils;
 import org.apache.ignite.internal.affinity.Assignment;
-import org.apache.ignite.internal.logger.IgniteLogger;
-import org.apache.ignite.internal.logger.Loggers;
 import org.apache.ignite.internal.metastorage.Entry;
 import org.apache.ignite.internal.metastorage.MetaStorageManager;
 import org.apache.ignite.internal.metastorage.WatchEvent;
-import org.apache.ignite.internal.metastorage.dsl.Condition;
 import org.apache.ignite.internal.metastorage.dsl.Iif;
 import org.apache.ignite.internal.metastorage.dsl.Operations;
 import org.apache.ignite.internal.table.distributed.replicator.TablePartitionId;
 import org.apache.ignite.internal.util.ByteUtils;
 import org.apache.ignite.lang.ByteArray;
-import org.jetbrains.annotations.NotNull;
 
 /**
  * Util class for methods needed for the rebalance process.
  */
+// TODO: https://issues.apache.org/jira/browse/IGNITE-18857 All rebalance logic and thus given RebalanceUtil should be moved to zones one.
 public class RebalanceUtil {
-
-    /** Logger. */
-    private static final IgniteLogger LOG = Loggers.forClass(RebalanceUtil.class);
-
-    /** Return code of metastore multi-invoke which identifies,
-     * that pending key was updated to new value (i.e. there is no active rebalance at the moment of call).
-     */
-    private static final int PENDING_KEY_UPDATED = 0;
-
-    /** Return code of metastore multi-invoke which identifies,
-     * that planned key was updated to new value (i.e. there is an active rebalance at the moment of call).
-     */
-    private static final int PLANNED_KEY_UPDATED = 1;
-
-    /** Return code of metastore multi-invoke which identifies,
-     * that planned key was removed, because current rebalance is already have the same target.
-     */
-    private static final int PLANNED_KEY_REMOVED_EQUALS_PENDING = 2;
-
-    /** Return code of metastore multi-invoke which identifies,
-     * that planned key was removed, because current assignment is empty.
-     */
-    private static final int PLANNED_KEY_REMOVED_EMPTY_PENDING = 3;
-
-    /** Return code of metastore multi-invoke which identifies,
-     * that assignments do not need to be updated.
-     */
-    private static final int ASSIGNMENT_NOT_UPDATED = 4;
-
-    /** Return code of metastore multi-invoke which identifies,
-     * that this trigger event was already processed by another node and must be skipped.
-     */
-    private static final int OUTDATED_UPDATE_RECEIVED = 5;
-
-    // TODO: sanpwc remove
-    /**
-     * Update keys that related to rebalance algorithm in Meta Storage. Keys are specific for partition.
-     *
-     * @param tableName Table name.
-     * @param partId Unique identifier of a partition.
-     * @param dataNodes Data nodes.
-     * @param replicas Number of replicas for a table.
-     * @param revision Revision of Meta Storage that is specific for the assignment update.
-     * @param metaStorageMgr Meta Storage manager.
-     * @param partNum Partition id.
-     * @param tableCfgPartAssignments Table configuration assignments.
-     * @return Future representing result of updating keys in {@code metaStorageMgr}
-     */
-    public static @NotNull CompletableFuture<Void> updatePendingAssignmentsKeys(
-            String tableName, TablePartitionId partId, Collection<String> dataNodes,
-            int replicas, long revision, MetaStorageManager metaStorageMgr, int partNum, Set<Assignment> tableCfgPartAssignments) {
-        ByteArray partChangeTriggerKey = partChangeTriggerKey(partId);
-
-        ByteArray partAssignmentsPendingKey = pendingPartAssignmentsKey(partId);
-
-        ByteArray partAssignmentsPlannedKey = plannedPartAssignmentsKey(partId);
-
-        ByteArray partAssignmentsStableKey = stablePartAssignmentsKey(partId);
-
-        Set<Assignment> partAssignments = AffinityUtils.calculateAssignmentForPartition(dataNodes, partNum, replicas);
-
-        boolean isNewAssignments = !tableCfgPartAssignments.equals(partAssignments);
-
-        byte[] partAssignmentsBytes = ByteUtils.toBytes(partAssignments);
-
-        //    if empty(partition.change.trigger.revision) || partition.change.trigger.revision < event.revision:
-        //        if empty(partition.assignments.pending)
-        //              && ((isNewAssignments && empty(partition.assignments.stable))
-        //                  || (partition.assignments.stable != calcPartAssighments() && !empty(partition.assignments.stable))):
-        //            partition.assignments.pending = calcPartAssignments()
-        //            partition.change.trigger.revision = event.revision
-        //        else:
-        //            if partition.assignments.pending != calcPartAssignments && !empty(partition.assignments.pending)
-        //                partition.assignments.planned = calcPartAssignments()
-        //                partition.change.trigger.revision = event.revision
-        //            else if partition.assignments.pending == calcPartAssignments
-        //                remove(partition.assignments.planned)
-        //                message after the metastorage invoke:
-        //                "Remove planned key because current pending key has the same value."
-        //            else if empty(partition.assignments.pending)
-        //                remove(partition.assignments.planned)
-        //                message after the metastorage invoke:
-        //                "Remove planned key because pending is empty and calculated assignments are equal to current assignments."
-        //    else:
-        //        skip
-
-        Condition newAssignmentsCondition;
-
-        if (isNewAssignments) {
-            newAssignmentsCondition = or(
-                    notExists(partAssignmentsStableKey),
-                    and(value(partAssignmentsStableKey).ne(partAssignmentsBytes), exists(partAssignmentsStableKey))
-            );
-        } else {
-            newAssignmentsCondition = and(value(partAssignmentsStableKey).ne(partAssignmentsBytes), exists(partAssignmentsStableKey));
-        }
-
-        var iif = iif(or(notExists(partChangeTriggerKey), value(partChangeTriggerKey).lt(ByteUtils.longToBytes(revision))),
-                iif(and(notExists(partAssignmentsPendingKey), newAssignmentsCondition),
-                        ops(
-                                put(partAssignmentsPendingKey, partAssignmentsBytes),
-                                put(partChangeTriggerKey, ByteUtils.longToBytes(revision))
-                        ).yield(PENDING_KEY_UPDATED),
-                        iif(and(value(partAssignmentsPendingKey).ne(partAssignmentsBytes), exists(partAssignmentsPendingKey)),
-                                ops(
-                                        put(partAssignmentsPlannedKey, partAssignmentsBytes),
-                                        put(partChangeTriggerKey, ByteUtils.longToBytes(revision))
-                                ).yield(PLANNED_KEY_UPDATED),
-                                iif(value(partAssignmentsPendingKey).eq(partAssignmentsBytes),
-                                        ops(remove(partAssignmentsPlannedKey)).yield(PLANNED_KEY_REMOVED_EQUALS_PENDING),
-                                        iif(notExists(partAssignmentsPendingKey),
-                                                ops(remove(partAssignmentsPlannedKey)).yield(PLANNED_KEY_REMOVED_EMPTY_PENDING),
-                                                ops().yield(ASSIGNMENT_NOT_UPDATED))
-                                ))),
-                ops().yield(OUTDATED_UPDATE_RECEIVED));
-
-        return metaStorageMgr.invoke(iif).thenAccept(sr -> {
-            switch (sr.getAsInt()) {
-                case PENDING_KEY_UPDATED:
-                    LOG.info(
-                            "Update metastore pending partitions key [key={}, partition={}, table={}, newVal={}]",
-                            partAssignmentsPendingKey.toString(), partNum, tableName,
-                            ByteUtils.fromBytes(partAssignmentsBytes));
-
-                    break;
-                case PLANNED_KEY_UPDATED:
-                    LOG.info(
-                            "Update metastore planned partitions key [key={}, partition={}, table={}, newVal={}]",
-                            partAssignmentsPlannedKey, partNum, tableName, ByteUtils.fromBytes(partAssignmentsBytes));
-
-                    break;
-                case PLANNED_KEY_REMOVED_EQUALS_PENDING:
-                    LOG.info(
-                            "Remove planned key because current pending key has the same value [key={}, partition={}, table={}, val={}]",
-                            partAssignmentsPlannedKey.toString(), partNum, tableName, ByteUtils.fromBytes(partAssignmentsBytes));
-
-                    break;
-                case PLANNED_KEY_REMOVED_EMPTY_PENDING:
-                    LOG.info(
-                            "Remove planned key because pending is empty and calculated assignments are equal to current assignments "
-                                    + "[key={}, partition={}, table={}, val={}]",
-                            partAssignmentsPlannedKey.toString(), partNum, tableName, ByteUtils.fromBytes(partAssignmentsBytes));
-
-                    break;
-                case ASSIGNMENT_NOT_UPDATED:
-                    LOG.debug(
-                            "Assignments are not updated [key={}, partition={}, table={}, val={}]",
-                            partAssignmentsPlannedKey.toString(), partNum, tableName, ByteUtils.fromBytes(partAssignmentsBytes));
-
-                    break;
-                case OUTDATED_UPDATE_RECEIVED:
-                    LOG.debug(
-                            "Received outdated rebalance trigger event [revision={}, partition={}, table={}]",
-                            revision, partNum, tableName);
-
-                    break;
-                default:
-                    throw new IllegalStateException("Unknown return code for rebalance metastore multi-invoke");
-            }
-        });
-    }
-
     /** Key prefix for pending assignments. */
     public static final String PENDING_ASSIGNMENTS_PREFIX = "assignments.pending.";
